+++ arviz/plots/backends/bokeh/khatplot.py
@@ -21,6 +21,7 @@ def plot_khat(
     figsize,
     xdata,
     khats,
+    good_k,
     kwargs,
     threshold,
     coord_labels,
@@ -53,7 +54,11 @@ def plot_khat(
 
     if hlines_kwargs is None:
         hlines_kwargs = {}
-    hlines_kwargs.setdefault("hlines", [0, 0.5, 0.7, 1])
+
+    if good_k is None:
+        good_k = 0.7
+
+    hlines_kwargs.setdefault("hlines", [0, good_k, 1])
 
     cmap = None
     if isinstance(color, str):
@@ -75,7 +80,7 @@ def plot_khat(
             rgba_c = cmap(color)
 
     khats = khats if isinstance(khats, np.ndarray) else khats.values.flatten()
-    alphas = 0.5 + 0.2 * (khats > 0.5) + 0.3 * (khats > 1)
+    alphas = 0.5 + 0.2 * (khats > good_k) + 0.3 * (khats > 1)
 
     rgba_c = vectorized_to_hex(rgba_c)
 
@@ -130,7 +135,7 @@ def plot_khat(
     xmax = len(khats)
 
     if show_bins:
-        bin_edges = np.array([ymin, 0.5, 0.7, 1, ymax])
+        bin_edges = np.array([ymin, good_k, 1, ymax])
         bin_edges = bin_edges[(bin_edges >= ymin) & (bin_edges <= ymax)]
         hist, _, _ = histogram(khats, bin_edges)
         for idx, count in enumerate(hist):